random_walk_1d: compute time-averaged MSD
For a random walk it makes sense to consider the time-averaged MSD which
greatly improves statistics.

Display the MSD on log-log scales as it "lives" on a logarithmic time
scale. Include a reference line with the expectation for long-time
diffusion.

Increase the "simulation" time and number of particles modestly, which
improves the statistics further.